More renaming fixes
